Add JavaScript Online
I'm adding javascript.onl to the list of exercises. It's a free collection of JS problems with an online judge. It can also be saved for offline use (everything is static). I explain a bit about the creation of it here: http://www.pesfandiar.com/blog/2016/05/12/javascript-online-hosting-static-site-cheaply

This is my first attempt at a pull request on Github, so please let me know if you'd like this in another format. Thanks